Update row colors to #83cdff in qtutils.cpp
Update qtutils.cpp to use #83cdff as row color for "isSelected" state (both light and dark mode).

Note: the color previously for "isSelected" was an 8 digit color, but when displayed, the code seems to use the last 6 digits of that 8 digit color code.  I changed it to a 6 digit color in this change.  The "IsHovered" is still an 8 digit color and the dispaly seems to also ignore the leading 2 digits of the color and only use the trailing 6 digits.  I did not change "isHovered" to a 6 digit color as I didn't want to change things that are not needed, but wanted to call this out so when someone sees the 8 digit color and tries to see what that is..the code actually ignore the first 2 digits when displaying the color, so it is confusing.